requirements.txt
python-dotenv==1.0.1
Faker==24.8.0
tqdm==4.66.4
